+++ cg/src/cg/input/records.cpp
@@ -26,8 +26,8 @@ std::string remark::write() const {
 std::vector<remark> remark::create(int number, const std::string &text) {
   auto max_width = 80 - 12 + 1;
   std::vector<remark> res;
-  for (int off = 0; off < text.size(); off += max_width) {
-    remark& rem = res.emplace_back();
+  for (int off = 0; off < (int)text.size(); off += max_width) {
+    remark &rem = res.emplace_back();
     rem.number = number;
     rem.text = text.substr(off, max_width);
   }
